refactor: move task demos to examples/ — not production code
ChessEvolveTask and SWEBenchTask are proof-of-concept demos showing
the Task contract works for both override and default run() paths.
They don't belong in the factory package.

- Move factory/tasks/chess_evolve.py → examples/chess_evolve_task.py
- Move factory/tasks/swe_bench.py → examples/swe_bench_task.py
- Remove factory/tasks/ directory
- Remove chess optional dependency from pyproject.toml
- Update test imports
